added metrics for service provider and method to prometheus handler
